requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
ytmusicapi>=1.5.0
python-dotenv==1.0.0 